<feat> create file CreateTable.jsx
