+++ src/App.jsx
@@ -1,7 +1,7 @@
 //import React, { } from 'react';
 import { createRoot } from 'react-dom/client'
 import Header from './assets/pages/Header';
-import Main from './assets/pages/Main';
+import { Main } from './assets/pages/Main';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 //* Estilos Primereact Component *//
@@ -10,7 +10,7 @@ import 'primereact/resources/primereact.min.css';
 import 'primeicons/primeicons.css';
 import 'primeflex/primeflex.css';
 
-function App() {
+export default function App() {
 
   return (
     <>
@@ -24,5 +24,3 @@ function App() {
 
 // Renderiza la aplicación directamente desde App.jsx
 createRoot(document.getElementById('root')).render(<App />);
-
-export default App;
